Add functions to "fonctions.py" file
And various modifications
